Rework tests
The goal was to get full MSI coverage, but infection seems have issues understanding what is covered
