+++ infection.json.dist
@@ -33,7 +33,7 @@
         "LogicalOrAllSubExprNegation": false,
         "LogicalOrNegation": false,
         "LogicalOrSingleSubExprNegation": false,
-        "PublicVisibility": true,
+        "PublicVisibility": false,
         "Throw_": false,
         "TrueValue": false,
         "UnwrapArrayKeys": false,
