+++ tests/CheckCorrectLogLevelTest.php
@@ -6,9 +6,8 @@ namespace WyriHaximus\Tests\PSR3;
 
 use PHPUnit\Framework\Attributes\DataProviderExternal;
 use PHPUnit\Framework\Attributes\Test;
-use PHPUnit\Framework\TestCase;
 use Psr\Log\InvalidArgumentException;
-use WyriHaximus\PSR3\Utils;
+use WyriHaximus\TestUtilities\TestCase;
 
 use function WyriHaximus\PSR3\checkCorrectLogLevel;
 
@@ -23,13 +22,6 @@ final class CheckCorrectLogLevelTest extends TestCase
     }
 
     #[Test]
-    #[DataProviderExternal(DataProvider::class, 'correctLogLevels')]
-    public function correctLogLevelUtils(string $logLevel): void
-    {
-        self::assertTrue(Utils::checkCorrectLogLevel($logLevel));
-    }
-
-    #[Test]
     #[DataProviderExternal(DataProvider::class, 'inCorrectLogLevels')]
     public function incorrectLogLevel(string $logLevel): void
     {
@@ -37,14 +29,5 @@ final class CheckCorrectLogLevelTest extends TestCase
 
         /** @phpstan-ignore function.deprecated */
         checkCorrectLogLevel($logLevel);
-    }
-
-    #[Test]
-    #[DataProviderExternal(DataProvider::class, 'inCorrectLogLevels')]
-    public function incorrectLogLevelUtils(string $logLevel): void
-    {
-        self::expectException(InvalidArgumentException::class);
-
-        Utils::checkCorrectLogLevel($logLevel);
     }
 }
