+++ tests/NormalizeContextTest.php
@@ -6,8 +6,7 @@ namespace WyriHaximus\Tests\PSR3;
 
 use PHPUnit\Framework\Attributes\DataProviderExternal;
 use PHPUnit\Framework\Attributes\Test;
-use PHPUnit\Framework\TestCase;
-use WyriHaximus\PSR3\Utils;
+use WyriHaximus\TestUtilities\TestCase;
 
 use function WyriHaximus\PSR3\normalizeContext;
 
@@ -23,16 +22,5 @@ final class NormalizeContextTest extends TestCase
     {
         /** @phpstan-ignore function.deprecated */
         self::assertSame($expectedOutput, normalizeContext($context));
-    }
-
-    /**
-     * @param array<string, mixed> $context
-     * @param array<string, mixed> $expectedOutput
-     */
-    #[Test]
-    #[DataProviderExternal(DataProvider::class, 'contexts')]
-    public function normalizeContextUtils(array $context, array $expectedOutput): void
-    {
-        self::assertSame($expectedOutput, Utils::normalizeContext($context));
     }
 }
